Move decoding logic into QueryStringEnumerable
